requirements.txt
# Основна бібліотека для створення веб-додатку
streamlit

# Для роботи з API Google Gemini (конспекти, питання)
google-generativeai

# Для завантаження відео/аудіо з YouTube
yt-dlp


# --- Бібліотеки для обробки файлів ---

# Для обробки PDF файлів (вилучення тексту)
pypdf2

# Для обробки PPTX файлів (вилучення тексту)
python-pptx